# controllers/customer.go
package controllers

import (
	"fmt"
	"net/http"
	"strconv"

	"github.com/ad3n/resto/models"
	"github.com/ad3n/resto/services"
	"github.com/gofiber/fiber/v2"
	"github.com/iancoleman/strcase"

	"github.com/go-playground/validator/v10"
)

type Customer struct {
	Service services.Customer
}

// @Description Get customer by ID
// @Summary Get customer by ID
// @Tags Customer
// @Accept json
// @Produce json
// @Param id path int true "Customer ID"
// @Success 200 {object} []models.Customer
// @Router /customers/{id} [get]
func (c Customer) Get(ctx *fiber.Ctx) error {
	cutomerId, err := strconv.Atoi(ctx.Params("id"))
	if err != nil {
		ctx.JSON(map[string]string{
			"message": "id is not number",
		})

		return ctx.SendStatus(http.StatusBadRequest)
	}

	customer, err := c.Service.Get(cutomerId)
	if err != nil {
		ctx.JSON(map[string]string{
			"message": "customer not found",
		})

		return ctx.SendStatus(fiber.StatusNotFound)
	}

	return ctx.JSON(customer)
}

// @Description Register new customer
// @Summary Register cew customer
// @Tags Customer
// @Accept json
// @Produce json
// @Param customer body forms.CustomerRegister true "Customer"
// @Success 200 {object} models.Customer
// @Router /customers [post]
func (c Customer) Register(ctx *fiber.Ctx) error {
	model := models.Customer{}
	ctx.BodyParser(&model)
	model.ID = 0

	validate := validator.New()
	err := validate.Struct(model)
	if err != nil {
		messages := []string{}
		for _, err := range err.(validator.ValidationErrors) {
			messages = append(messages, fmt.Sprintf("%s is %s", strcase.ToSnake(err.Field()), err.Tag()))
		}

		ctx.JSON(map[string]interface{}{
			"message": messages,
		})

		return ctx.SendStatus(fiber.StatusBadRequest)
	}

	model, err = c.Service.Save(model)
	if err != nil {
		ctx.JSON(map[string]string{
			"message": err.Error(),
		})

		return ctx.SendStatus(fiber.StatusInternalServerError)
	}

	ctx.JSON(model)

	return ctx.SendStatus(fiber.StatusCreated)
}

// @Description Customer reservation
// @Summary Customer Reservation
// @Tags Customer
// @Accept json
// @Produce json
// @Param id path int true "Customer ID"
// @Param reservation body forms.CustomerReservation true "Customer"
// @Success 200 {object} models.Order
// @Router /customers/{id}/reservation [post]
func (c Customer) Reservation(ctx *fiber.Ctx) error {
	customerId, err := strconv.Atoi(ctx.Params("id"))
	if err != nil {
		ctx.JSON(map[string]string{
			"message": "id is not number",
		})

		return ctx.SendStatus(http.StatusBadRequest)
	}

	customer, err := c.Service.Get(customerId)
	if err != nil {
		ctx.JSON(map[string]string{
			"message": err.Error(),
		})

		return ctx.SendStatus(http.StatusNotFound)
	}

	model := models.Order{}
	ctx.BodyParser(&model)

	validate := validator.New()
	err = validate.Struct(model)
	if err != nil {
		messages := []string{}
		for _, err := range err.(validator.ValidationErrors) {
			messages = append(messages, fmt.Sprintf("%s is %s", strcase.ToSnake(err.Field()), err.Tag()))
		}

		ctx.JSON(map[string]interface{}{
			"message": messages,
		})

		return ctx.SendStatus(fiber.StatusBadRequest)
	}

	model, err = c.Service.Reservation(customer, model.TableNumber)
	if err != nil {
		ctx.JSON(map[string]string{
			"message": err.Error(),
		})

		return ctx.SendStatus(fiber.StatusInternalServerError)
	}

	ctx.JSON(model)

	return ctx.SendStatus(fiber.StatusCreated)
}